fix(menu): se gestiona menú desde service
